Flow UART pipeline: accumulate counts, don't sample the latest frame
Root cause of "same move, random magnitude" (X read ~8% of true displacement
varying run-to-run, post-stop creep, wrong-sign blips on direction change):

- uart_read_bytes requested 256 bytes with a 100ms timeout; at 19200 baud
  that's a 133ms fill, so the timeout fired every time and frames were
  parsed in ~100ms batches of ~21.
- get_data's latest-frame-wins semantics dropped every frame but the last
  of each batch: ~95% of displacement counts silently discarded, and the
  consumed motion lagged reality by up to 100ms (the creep after stopping).

Fix: 10ms UART timeout (<=2 frames per batch) and the driver now accumulates
acc_x/acc_y which get_data hands over and zeroes; flow_hold_update converts
the accumulated counts over the measured consume window (FRAME_DT_MAX 0.2s).
Exact regardless of batching. The pre-metric position feedback used the
driver-side integral and was lossless - the metric refactor introduced this
regression by deriving velocity from per-frame counts via the lossy channel.

+++ components/control/flow_hold.h
@@ -53,11 +53,12 @@ void flow_hold_set_flow_scale(flow_hold_t *fh, float scale);
 void flow_hold_predict(flow_hold_t *fh, float ax_world, float ay_world, float dt);
 
 /**
- * @brief 有新光流帧时调用（~50-100Hz）：陀螺补偿（counts 域）→ 按 TOF 高度换算
- *        米制速度 v = counts × flow_scale × height / dt_frame → 互补滤波校正 vx_est。
- *        米制化后环路增益不再随悬停高度漂移（这是定点不稳/起飞漂移的主根因）。
+ * @brief 有新光流数据时调用：flow_x/flow_y 传驱动的**累计** counts（acc_x/acc_y，
+ *        自上次消费以来的总位移——不要传最新帧值，批量到达会丢帧）。
+ *        换算 v = 累计counts × flow_scale × height / 消费窗口时长（实测），
+ *        再做米制域陀螺补偿与互补滤波校正 vx_est。
  */
-void flow_hold_update(flow_hold_t *fh, int16_t flow_x, int16_t flow_y,
+void flow_hold_update(flow_hold_t *fh, float flow_x, float flow_y,
                       float gyro_x, float gyro_y, uint8_t qual, float height_m);
 
 void flow_hold_reset(flow_hold_t *fh);
